Fix bindings not working on raid frames
This commit fixes an issue where using bindings on raid frames were not
working at all.

The root cause of the issue was that raid frames were not included, nor
updated to the internal table of supported frames. Since raid frames,
specifically the compact raid frames, are dynamic and added to the UI
on-demand, active raid frames must be re-collected again every time
there's a change in the group's or raid's roster (incl. pets).

To implement this, I created a handler that collects active raid frames
and maintains the addon's internal `self.ccframes` table by always
keeping the default frames while refreshing the active raid frames and
pruning the staled ones.

The created handler reacts to messages returned by the following events:
* `PLAYER_ENTERING_WORLD`
* `RAID_ROSTER_UPDATE`
* `PARTY_MEMBERS_CHANGED`
* `UNIT_PET`
* `CVAR_UPDATE`